Changed ImGui provider, moved more into components, split apart the vendor libraries and more

+++ SoulEngine/Props/Properties/DoubleProperty.cs
@@ -1,4 +1,4 @@
-using ImGuiNET;
+using Hexa.NET.ImGui;
 using SoulEngine.Data.NBT;
 
 namespace SoulEngine.Props;
@@ -13,7 +13,7 @@ public class DoubleProperty : SerializedProperty<double>
     public override unsafe void Edit()
     {
         fixed (void* ptr = &Value)
-            ImGui.InputScalar(Name, ImGuiDataType.Double, (IntPtr)ptr);
+            ImGui.InputScalar(Name, ImGuiDataType.Double, ptr);
     }
